refactor: comprehensive code cleanup and optimization pass
- Remove unused imports and dead code across all modules
- Apply clippy suggestions for better Rust idioms:
  * Use div_ceil() instead of manual ceiling division
  * Use fill(0) and is_empty() for cleaner code
  * Prefer iterators over manual indexing where possible
  * Simplify lifetime annotations and remove unnecessary casts
  * Use rotate_left() and other optimized bit operations
- Streamline error handling with consistent conversion patterns
- Remove internal-only constructors and unused helper functions
- Optimize memory operations and reduce unnecessary allocations
- Improve constant-time operations in cryptographic code
- Consolidate similar code patterns across modules

This cleanup improves code maintainability, performance, and adherence
to Rust best practices while maintaining API compatibility.

+++ crates/algorithms/src/ec/p521/point.rs
@@ -342,7 +342,7 @@ impl Point {
         // Montgomery ladder — MSB→LSB
         for byte in scalar.as_secret_buffer().as_ref() {
             for bit in (0..8).rev() {
-                let choice = Choice::from(((byte >> bit) & 1) as u8);
+                let choice = Choice::from((byte >> bit) & 1);
 
                 ProjectivePoint::conditional_swap(&mut r0, &mut r1, choice);
                 // "ladder step"
